+++ src/JustSending/Data/IDataStore.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Text.Json;
 using System.Threading.Tasks;
-using JustSending.Services;
 
 namespace JustSending.Data
 {
