+++ src/JustSending/Data/JsonShortNamePolicy.cs
@@ -1,6 +1,4 @@
 using System.Collections.Generic;
-using System.Security.Cryptography;
-using System.Text;
 using System.Text.Json;
 using JustSending.Services;
