+++ src/JustSending/Views/Shared/_Layout.cshtml
@@ -56,9 +56,9 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/sweetalert/1.1.3/sweetalert.min.js" integrity="sha256-egVvxkq6UBCQyKzRBrDHu8miZ5FOaVrjSqQqauKglKc=" crossorigin="anonymous"></script>
     <script src="https://browser.sentry-cdn.com/5.17.0/bundle.min.js" crossorigin="anonymous"></script>
     *@  
-    <script type="text/javascript" src="~/js/combined-main.min.js" asp-append-version="true" integrity="sha384-iqcP4m3MyNLBHZBx9svGfh+Mk1oy/uRl8WORi1Os8kMM0oHSHUAEYSImJAsJmE+6" crossorigin="anonymous"></script>
+    <script type="text/javascript" src="~/js/combined-main.min.js" asp-append-version="true" integrity="sha384-UzoaLAKICBSbjnbswZQwFTzEPohV9KHfeefe1Uen0khfkmj8mwFd5j57m/LgPAuT" crossorigin="anonymous"></script>
 
-    <script type="text/javascript">
+<script type="text/javascript">
     if (location.href.indexOf("localhost") === -1) {
         Sentry.init({ dsn: '@Environment.GetEnvironmentVariable("Sentry__Dsn")' });
     }
