--- test/load-testing/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "load-testing",
-  "version": "1.0.0",
-  "description": "",
-  "main": "basic.js",
-  "scripts": {
-    "start": "node basic.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "npm": "^6.14.5",
-    "puppeteer": "^4.0.0",
-    "random": "^2.2.0"
-  }
-}
